landingPage reorganized
