+++ components/listingsPage/index.js
@@ -1,7 +1,7 @@
 import React, { Fragment } from "react";
 
-import Filters from "./filters";
-import Results from "./results";
+import Filters from "./Filters";
+import Results from "./Results";
 import { Row, Col, Heading } from "@moneypensionservice/directories";
 
 const Listing = ({ t }) => {
